Add snapshot tests for rendering and equipping each goody
This will help highlight changes to either how the goodies are packaged, merged, or rendered.
