+++ tools/adventure-pack/src/app/constants.tsx
@@ -1,0 +1,9 @@
+import type { Language } from "./Language";
+
+export const LANGUAGE_NAMES: Record<Language, string> = {
+  java: "Java",
+  javascript: "JavaScript",
+  kotlin: "Kotlin",
+  python3: "Python 3",
+  typescript: "TypeScript",
+};
